// angularapp/src/main/angularapp/src/app/openapi/model/product.ts
/**
 * Till API
 * This app provides REST APIs for all associated till calls. From creating pages, products to transaction
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Group } from './group';

export interface Product { 
    id?: number;
    group?: Group;
    name?: string;
    info?: string;
    price?: number;
    image?: string;
}