// angularapp/src/main/angularapp/src/app/openapi/model/transactionRecord.ts
/**
 * Till API
 * This app provides REST APIs for all associated till calls. From creating pages, products to transaction
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Product } from './product';

export interface TransactionRecord { 
    product?: Product;
    quantity?: number;
    price?: number;
}